# Documents/MayDay_REST_API_Documentation.yaml
swagger: '2.0'
info:
  title: SSNoC MayDay API
  description: 'SSNoC MayDay API_UserCase:Join Community Chat Publicly'
  version: 1.0.2
host: '172.29.93.226:8080'
schemes:
  - http post
basePath: /
produces:
  - application/json
  - application/html
paths:
...
  '/users/onlineUsers':
    get:
      description: |
        fetch online user list, http method is GET
      tags:
        - Users
      responses:
        '200':
          description: 'online user list found'
          schema:
            name: users
            type: array
        '500':
          description: 'online user list not found!'
          schema:
            type: array
          items:
            $ref: '#/definitions/Error'
  '/users/user/:user':
    get:
      description: |
        Fetch a Single User, http method is GET
      tags:
        - Users
      responses:
        '200':
          description: 'Fetched a Single User'
          schema:
            name: user
            type: array
        '500':
          description: 'fail to fetch a Single User!'
          schema:
            type: array
          items:
            $ref: '#/definitions/Error'
  '/users/updateUser/:uid':
    put:
      description: update user
      tags:
        - Users
      responses:
        '200':
          description: 'User updated successfully'
          schema:
            type: array
        '404':
          description: 'User update failed!'
          schema:
            type: array
          items:
            $ref: '#/definitions/Error'
  '/publicmessage/wall':
    get:
      description: get historic chat messages
      tags:
        - Public Chat
      responses:
        '200':
          description: 'Messages found'
          schema:
            type: array
            item:
              $ref: '#/definitions/Activity'
        '500':
          description: 'Error no messages found!'
          schema:
            type: array
          items:
            $ref: '#/definitions/Error'
  '/publicmessage/userConnected/:user':
    get:
      description: new user enters public chatroom
      tags:
        - Public Chat
      responses:
        '200':
          description: 'get online users'
          schema:
            type: array
            item:
              $ref: '#/definitions/Activity'
        '500':
          description: 'online user not found!'
          schema:
            type: array
          items:
            $ref: '#/definitions/Error'
  '/publicmessage/chatScrollUp/:msgid':
    get:
      description: Scroll up for more messages
      tags:
        - Public Chat
      responses:
        '200':
          description: 'get more historic messages'
          schema:
            type: array
            item:
              $ref: '#/definitions/Activity'
        '500':
          description: 'historic messages not found'
          schema:
            type: array
          items:
            $ref: '#/definitions/Error'
  '/publicmessage/message':
    post:
      description: post a new message
      tags:
        - Public Chat
      responses:
        '201':
          description: 'Message posted!'
          schema:
            type: array
            item:
              $ref: '#/definitions/Activity'
        '404':
          description: 'User does not exist!'
          schema:
            type: array
          items:
            $ref: '#/definitions/Error'
...
  '/users/logout/:user':
    get:
      description: User logged out
      tags:
        - Join Community
      responses:
        '200':
          description: Redirect page to login
          schema:
            type: string
        '500':
          description: Logout unsucessful
          schema:
            type: array
          items:
            $ref: '#/definitions/Error'